pyproject.toml
[project]
name = "adk-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
